requirements.txt
# Web3 and Blockchain
web3==6.15.1
etherscan-python==2.1.0
python-decouple==3.8

# Data Collection
requests==2.31.0
aiohttp==3.9.3
yfinance==0.2.18
fredapi==0.5.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Database (simplified for testing)
sqlalchemy==2.0.25

# Machine Learning (basic)
scikit-learn==1.4.0
statsmodels==0.14.1

# Visualization
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2

# Development
pytest==8.0.0
jupyter==1.0.0

# Network Analysis
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.2